Refactor library to remove tf support and simplify code
